+++ .github/envs/environment.yml
@@ -2,6 +2,7 @@ name: test-environment
 channels:
   - conda-forge
 dependencies:
+  - python ==3.12
   - dask-core
   - python-duckdb
   - pandas
